Fixes the include path to find the correct model
If you don't specify the prefix param it won't find the correct model in the frontend while trying to edit a form.
Note that on line 44 it specifies the prefix on the getInstance, and that prefix will be used as key on the paths array. Without this fix, it seems like the line 43 is useless since it will add a path to somewhere it won't be found.