Update element.php
'object' (which is in the request) is lost when paginating through items (actually, any time there is a refresh in the modal window), resulting in the item selected not being picked up when properly coding an 'item' custom field to allow for multiple occurrences on a page. When there is just one 'item' custom field (original from core K2), the bug gets unnoticed (since 'object' is never used in jSelectItem of the 'item' custom field).