Allow YOOtheme Pro integration
See https://www.joomlaworks.net/forum/k2-en/64209-fotis-and-k2-devs-k2-as-yoothemepro-source#179948

I've added a check to see if we're in YTP to trigger the event, but IMHO the overhead of triggering an event like this is way less than doing the check, 
because it may lead to a load of the template there. Feel free to remove it.