Added new "Description" option in K2 tags
This option allows for displaying a custom description text in tag listings (right below the tag name - needs to be uncommented in the tag.php sub-template override for now - a display option will be available in an upcoming update) as well as the relevant meta description (generic, opengraph and twitter). This is to assist in customizing your SEO for tag listings.